requirements.txt
langchain
langchain_openai
openai
python-dotenv
chromadb
streamlit
streamlit_chat
pypdf